feat: Add admin actions to film and restaurant index views

+++ views/film/index.php
@@ -1,7 +1,10 @@
 <?php include 'views/includes/header.php'; ?>
 
 <h1>Films</h1>
-<a href="index.php?controller=film&action=add" class="btn">Ajouter un Film</a>
+<?php if (AuthController::checkAdmin()): ?>
+  <a href="index.php?controller=film&action=add" class="btn">Ajouter un Film</a>
+<?php endif; ?>
+
 <table>
   <tr>
     <th>ID</th>
@@ -11,7 +14,9 @@
     <th>Réalisateur</th>
     <th>Langue</th>
     <th>Année</th>
-    <th>Actions</th>
+    <?php if (AuthController::checkAdmin()): ?>
+      <th>Actions</th>
+    <?php endif; ?>
   </tr>
   <?php foreach ($films as $film): ?>
   <tr>
@@ -26,10 +31,13 @@
     <td><?= $film->realisateur; ?></td>
     <td><?= $film->langue; ?></td>
     <td><?= $film->annee; ?></td>
-    <td class="action-buttons">
-      <a href="index.php?controller=film&action=edit&id=<?= $film->id; ?>" class="edit"><i class="fas fa-edit"></i> Modifier</a>
-      <a href="index.php?controller=film&action=delete&id=<?= $film->id; ?>" class="delete"><i class="fas fa-trash-alt"></i> Supprimer</a>
-    </td>
+    <?php if (AuthController::checkAdmin()): ?>
+      <td class="action-buttons">
+        <a href="index.php?controller=film&action=edit&id=<?= $film->id; ?>" class="edit"><i class="fas fa-edit"></i> Modifier</a>
+        <a href="index.php?controller=film&action=delete&id=<?= $film->id; ?>" class="delete"><i class="fas fa-trash-alt"></i> Supprimer</a>
+      </td>
+    <?php endif; ?>
+
   </tr>
   <?php endforeach; ?>
 </table>
