+++ views/restaurant/index.php
@@ -1,14 +1,19 @@
 <?php include 'views/includes/header.php'; ?>
 
 <h1>Restaurants</h1>
-<a href="index.php?controller=restaurant&action=add" class="btn">Ajouter un Restaurant</a>
+<?php if (AuthController::checkAdmin()): ?>
+    <a href="index.php?controller=restaurant&action=add" class="btn">Ajouter un Restaurant</a>
+<?php endif; ?>
+
 <table>
     <tr>
         <th>ID</th>
         <th>Nom</th>
         <th>Adresse</th>
         <th>Restaurateur</th>
-        <th>Actions</th>
+        <?php if (AuthController::checkAdmin()): ?>
+            <th>Actions</th>
+        <?php endif; ?>
     </tr>
     <?php foreach ($restaurants as $restaurant): ?>
     <tr>
@@ -16,10 +21,13 @@
         <td><?php echo $restaurant->nom; ?></td>
         <td><?php echo $restaurant->adresse; ?></td>
         <td><?php echo $restaurant->restaurateur; ?></td>
-        <td class="action-buttons">
+        <?php if (AuthController::checkAdmin()): ?>
+          <td class="action-buttons">
             <a href="index.php?controller=restaurant&action=edit&id=<?php echo $restaurant->id; ?>" class="edit"><i class="fas fa-edit"></i> Modifier</a>
             <a href="index.php?controller=restaurant&action=delete&id=<?php echo $restaurant->id; ?>" class="delete"><i class="fas fa-trash-alt"></i> Supprimer</a>
-        </td>
+          </td>
+        <?php endif; ?>
+        
     </tr>
     <?php endforeach; ?>
 </table>
